tests/wpt/web-platform-tests/websockets/keeping-connection-open/001.html

<!doctype html>
<title>WebSockets: 20s inactivity after handshake</title>
<meta name=timeout content=long>
<p>Note: This test takes 20 seconds to run.</p>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script src=../constants.js?pipe=sub></script>
<meta name="variant" content="">
<meta name="variant" content="?wss">
<div id=log></div>
<script>
async_test(function(t) {
  var ws = new WebSocket(SCHEME_DOMAIN_PORT+'/echo');
  ws.onclose = ws.onerror = ws.onmessage = t.unreached_func();
  ws.onopen = t.step_func(function(e) {
    t.step_timeout(function() {
      ws.send('test');
      ws.onmessage = t.step_func(function(e) {
        assert_equals(e.data, 'test');
        ws.onclose = t.step_func(function(e) {
          t.step_timeout(() => t.done(), 50);
        });
        ws.close();
      });
    }, 20000);
  })
}, null, {timeout:30000});
</script>
